Optimized implementation to order, cancel and delete an order

+++ order-service/src/main/java/com/order/processing/system/order/service/services/impl/OrderServiceImpl.java
@@ -4,14 +4,22 @@ package com.order.processing.system.order.service.services.impl;
 import com.order.processing.system.order.service.dto.request.CancelOrderRequestDTO;
 import com.order.processing.system.order.service.dto.request.ConfirmOrderRequestDTO;
 import com.order.processing.system.order.service.dto.request.CreateOrderRequestDTO;
+import com.order.processing.system.order.service.dto.request.UpdateProductStockRequestDTO;
 import com.order.processing.system.order.service.dto.response.ApiResponse;
 import com.order.processing.system.order.service.dto.response.ProductResponseDTO;
 import com.order.processing.system.order.service.enums.OrderStatus;
+import com.order.processing.system.order.service.exceptions.BadRequestException;
 import com.order.processing.system.order.service.exceptions.RecordNotFoundException;
 import com.order.processing.system.order.service.model.Order;
 import com.order.processing.system.order.service.repository.OrderRepository;
 import com.order.processing.system.order.service.services.OrderService;
 import com.order.processing.system.order.service.util.Mapper;
+import com.order.processing.system.order.service.util.Utils;
+import com.order.processing.system.proto.CheckStockRequest;
+import com.order.processing.system.proto.ProductResponse;
+import com.order.processing.system.proto.ProductServiceGrpc;
+import com.order.processing.system.proto.ViewAllProductsRequest;
+import io.grpc.StatusRuntimeException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -22,6 +30,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
@@ -34,6 +43,8 @@ import static com.order.processing.system.order.service.util.AppMessages.*;
 @Slf4j
 public class OrderServiceImpl implements OrderService {
     private final OrderRepository orderRepository;
+    private final ProductServiceGrpc.ProductServiceBlockingStub productServiceBlockingStub;
+    private final ProductGrpcService productService;
 
     @Override
     public ResponseEntity<ApiResponse> createOrder(CreateOrderRequestDTO createOrderRequestDTO) {
@@ -44,8 +55,12 @@ public class OrderServiceImpl implements OrderService {
         mappedOrder.setStatus(OrderStatus.PENDING);
         Order savedOrder = orderRepository.save(mappedOrder);
 
-        //TODO call inventory service to subtract the quantity of order placed from the stock
-
+        //Make call inventory service to subtract the quantity of order placed from the stock
+        productService.updateStockAvailability(
+                UpdateProductStockRequestDTO.builder()
+                        .productUuid(createOrderRequestDTO.getProductUuid())
+                        .stockAvailable(productResponse.getQuantity() - createOrderRequestDTO.getQuantity())
+                        .build());
         return ResponseEntity.status(HttpStatus.CREATED).body(
                 new ApiResponse<>(true,
                         HttpStatus.CREATED.value(),
@@ -56,9 +71,30 @@ public class OrderServiceImpl implements OrderService {
     }
 
     private ProductResponseDTO checkProductAvailabilityInInventoryService(CreateOrderRequestDTO createOrderRequestDTO) {
-
-        return ProductResponseDTO.builder()
+        CheckStockRequest request = CheckStockRequest.newBuilder()
+                .setProductUuid(createOrderRequestDTO.getProductUuid())
                 .build();
+        try {
+            com.order.processing.system.proto.ApiResponse response = this.productServiceBlockingStub.checkStockAvailability(request);
+            if (response.getSuccess()) {
+                ProductResponse productResponse = response.getData(0);
+                if (productResponse.getQuantity() < createOrderRequestDTO.getQuantity()) {
+                    throw new BadRequestException("Oops! The quantity requested is less than the quantity available in stock (" + productResponse.getQuantity() + ")");
+                } else {
+                    return ProductResponseDTO.builder()
+                            .uuid(productResponse.getUuid())
+                            .name(productResponse.getName())
+                            .quantity(productResponse.getQuantity())
+                            .unitPrice(productResponse.getUnitPrice())
+                            .dateCreated(Utils.toLocalDateTime(productResponse.getDateCreated()))
+                            .lastModified(Utils.toLocalDateTime(productResponse.getLastModified()))
+                            .build();
+                }
+            }
+            throw new BadRequestException(NO_PRODUCT_FOUND_WITH_ID);
+        } catch (StatusRuntimeException e) {
+            throw new BadRequestException(COULD_NOT_PROCESS_REQUEST);
+        }
     }
 
     @Override
@@ -81,12 +117,40 @@ public class OrderServiceImpl implements OrderService {
     public ResponseEntity<ApiResponse> cancelOrder(CancelOrderRequestDTO cancelOrderRequestDTO) {
         Optional<Order> optional = validateOrderUuid(cancelOrderRequestDTO.getOrderUuid());
         Order order = optional.get();
+        if (OrderStatus.SUCCESS.equals(order.getStatus())) {
+            throw new BadRequestException(SUCCESSFUL_ORDER_CANNOT_BE_CANCELED);
+        }
+        if (OrderStatus.CANCELLED.equals(order.getStatus())) {
+            throw new BadRequestException(ORDER_ALREADY_CANCELED);
+        }
         order.setStatus(OrderStatus.CANCELLED);
         order.setLastModified(LocalDateTime.now());
         order = orderRepository.save(order);
-        //TODO Add back the quantity of the product in inventory service for the order that was initially reserved
-        return ResponseEntity.status(HttpStatus.OK).body(
-                new ApiResponse<>(true,
+
+        //Add back the quantity of the product in inventory service for the order that was initially reserved
+        ViewAllProductsRequest request = ViewAllProductsRequest.newBuilder()
+                .setPage(0)
+                .setSize(1000)
+                .build();
+        com.order.processing.system.proto.ApiResponse response = this.productServiceBlockingStub.viewAllProducts(request);
+        if (response.getSuccess()) {
+            List<ProductResponse> productResponseList = response.getDataList();
+            if (!productResponseList.isEmpty()) {
+                Order finalOrder = order;
+                ProductResponse response1 = productResponseList.stream()
+                        .filter(productResponse -> productResponse.getName().equalsIgnoreCase(finalOrder.getProduct()))
+                        .findFirst()
+                        .get();
+                productService.updateStockAvailability(
+                        UpdateProductStockRequestDTO.builder()
+                                .productUuid(response1.getUuid())
+                                .stockAvailable(response1.getQuantity() + finalOrder.getQuantity())
+                                .build());
+            }
+        }
+
+        return ResponseEntity.status(HttpStatus.OK).
+                body(new ApiResponse<>(true,
                         HttpStatus.OK.value(),
                         HttpStatus.OK,
                         ORDER_CANCELLED_SUCCESSFULLY,
@@ -139,6 +203,9 @@ public class OrderServiceImpl implements OrderService {
     @Override
     public ResponseEntity<ApiResponse> deleteOrder(String orderUuid) {
         Optional<Order> optional = validateOrderUuid(orderUuid);
+        if (OrderStatus.PENDING.equals(optional.get().getStatus())) {
+            throw new BadRequestException(PENDING_ORDER_CANNOT_BE_DELETED);
+        }
         orderRepository.deleteOrderByUuid(optional.get().getUuid());
         return ResponseEntity.status(HttpStatus.OK).body(
                 new ApiResponse<>(true,
